requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.1
django-better-admin-arrayfield==1.4.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==22.1
django-ipware==4.0.2
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.4
django-mptt==0.14.0
django-recaptcha==3.0.0
django-resized==1.0.2
djangorestframework==3.14.0
drf-recaptcha==2.2.1
drf-yasg==1.21.4
filelock==3.8.0
flake8==5.0.4
flake8-django==1.1.5
identify==2.5.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.31
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-environ==0.4.54
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
wrapt==1.14.1
django-rest-multiple-models==2.1.3